+++ nncf/tensor/functions/torch_numeric.py
@@ -9,12 +9,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import Any, Callable, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
-from safetensors.torch import load_file as pt_load_file
-from safetensors.torch import save_file as pt_save_file
 
 from nncf.tensor import TensorDataType
 from nncf.tensor import TensorDeviceType
@@ -467,18 +465,3 @@ def _(a: torch.Tensor) -> torch.Tensor:
 @numeric.ceil.register(torch.Tensor)
 def _(a: torch.Tensor) -> torch.Tensor:
     return torch.ceil(a)
-
-
-def load_file(
-    file_path: str,
-    *,
-    device: Optional[TensorDeviceType] = None,
-) -> Dict[str, torch.Tensor]:
-    if device is not None:
-        device = DEVICE_MAP[device]
-    return pt_load_file(file_path, device=device)
-
-
-@numeric.save_file.register(torch.Tensor)
-def _(data: Dict[str, torch.Tensor], file_path: str) -> None:
-    return pt_save_file(data, file_path)
